+++ test/dummy/config/routes.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
 Rails.application.routes.draw do
-  mount Ideasbugs::Engine => "/feedback"
+  mount_ideasbugs at: "/feedback"
   get "sample", to: "sample#show"
 end